requirements.txt
discord.py
requests
python-bitget
websocket-client
python-dotenv

